backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Autonomous Landing Bay RL Environment Backend"
authors = [
    {name = "lucdev", email = "me@lucdev.net"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.0",
    "uvicorn[standard]>=0.38.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.44",
    "gymnasium>=1.2.2",
    "numpy>=2.0.0",
    "python-multipart>=0.0.20",
    "websockets>=15.0.1",
    "torch>=2.0.0",
    "stable-baselines3>=2.3.0",
    "pydantic[email] (>=2.12.3,<3.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "questionary (>=2.1.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poe.tasks]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "poethepoet (>=0.37.0,<0.38.0)"
]
